+++ R/runSimulation.R
@@ -457,14 +457,14 @@
 #'   \code{debug = 'A1'} will debug only the first function in this list, and all remaining analysis
 #'   functions will be ignored.
 #'
-#'   For debugging specific rows in the \code{Design} input the specific row number can be appended to the
-#'   \code{debug} input to obtain this specific subset using a \code{'-'} separator.
-#'   For instance, debugging the \code{analyse}
-#'   function in the second row of \code{Design} can be declared via \code{debug = 'analyse-2'}. This feature
-#'   is included to avoid taking manual subsets of the \code{Design} object, however this still remains a
-#'   viable approach.
+#'   For debugging specific rows in the \code{Design} input (e.g.,
+#'   when a number of initial rows successfully complete but the \code{k}th
+#'   row fails) the row number can be appended to the standard
+#'   \code{debug} input using a \code{'-'} separator.
+#'   For instance, debugging whenever an error is raised
+#'   in the second row of \code{Design} can be declared via \code{debug = 'error-2'}.
 #'
-#'   Alternatively, users may place \code{\link{browser}} calls within the respective functions for
+#'   Finally, users may place \code{\link{browser}} calls within the respective functions for
 #'   debugging at specific lines, which is useful when debugging based on conditional evaluations (e.g.,
 #'   \code{if(this == 'problem') browser()}). Note that parallel computation flags
 #'   will automatically be disabled when a \code{browser()} is detected or when a debugging
@@ -870,11 +870,17 @@ runSimulation <- function(design, replications, generate, analyse, summarise,
         useFuture <- tolower(parallel) == 'future'
         parallel <- TRUE
     } else useFuture <- FALSE
+    if(is.null(seed)){
+        seed <- if(missing(design))
+            rint(1L, min=1L, max = 2147483647L)
+        else rint(nrow(design), min=1L, max = 2147483647L)
+    }
     if(debug != 'none'){
         if(grepl('-', debug)){
             tmp <- strsplit(debug, '-')[[1]]
             debug <- tmp[1L]
-            Design <- Design[as.numeric(tmp[2L]), , drop=FALSE]
+            design <- design[as.integer(tmp[2L]), , drop=FALSE]
+            seed <- seed[as.integer(tmp[2L])]
         }
     }
     if(is.list(analyse)){
@@ -902,11 +908,6 @@ runSimulation <- function(design, replications, generate, analyse, summarise,
                 }
             }
         }
-    }
-    if(is.null(seed)){
-        seed <- if(missing(design))
-            rint(1L, min=1L, max = 2147483647L)
-        else rint(nrow(design), min=1L, max = 2147483647L)
     }
     stopifnot(notification %in% c('none', 'condition', 'complete'))
     if(notification != 'none')